--- db/__init__.py
@@ -1,1 +1,0 @@
-"""Database container and migration helpers."""
